[Infra] Fix new code analysis warnings
- Use pattern matching.
- Remove redundant cast.

+++ src/OpenTelemetry.Configuration.Declarative/Internal/Yaml/YamlScalarConverter.cs
@@ -106,7 +106,7 @@ internal static class YamlScalarConverter
             accumulator = (accumulator * numberBase) + digit;
         }
 
-        return accumulator > (ulong)long.MaxValue
+        return accumulator > long.MaxValue
             ? ConfigValue.UnrepresentableInteger()
             : ConfigValue.Integer((long)accumulator);
     }
